[XB1] Refactor to remove repeated code
Full change including internal files at http://go/cobalt-cl/250460

b/277971529
